.blackbox5/2-engine/01-core/infrastructure/requirements.txt
# Black Box 5 Engine - Python Dependencies

# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# CLI
click>=8.1.0

# Configuration
pyyaml>=6.0.1

# Async Support
asyncio-mqtt>=0.16.0

# Health Monitoring
psutil>=5.9.0

# WebSockets
websockets>=12.0

# HTTP Client
httpx>=0.25.0
aiohttp>=3.9.0

# Data Processing
python-dateutil>=2.8.2

# Logging
python-json-logger>=2.0.7

# Development Dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.11.0
isort>=5.12.0
mypy>=1.7.0

# Optional: Brain/RAG System (Phase 3)
# chromadb>=0.4.0
# sentence-transformers>=2.2.0
# neo4j>=5.14.0

# Optional: MCP Integrations (Phase 3)
# mcp>=0.1.0
